+++ include/BasePhysics/Core/Geometry.hpp
@@ -10,7 +10,7 @@
 namespace bp::geometry
 {
     float PointSegmentDistance(Vec2 point, Vec2 segmentA, Vec2 segmentB, Vec2 &outClosestPoint);
-    int FindClosestPointIndex(Vec2 center, const std::vector<Vec2> &vertices);
+    size_t FindClosestPointIndex(Vec2 center, const std::vector<Vec2> &vertices);
     void ProjectVertices(const std::vector<Vec2> &vertices, Vec2 axis, float &outMin, float &outMax);
     void ProjectCircle(Vec2 center, float radius, Vec2 axis, float &outMin, float &outMax);
 }
